+++ tools/wheels/cibw_test_command.sh
@@ -7,20 +7,29 @@ cd $PROJECT_DIR/tests
 #HACK: if the tests are not split up the CI runs out of memory...
 
 # test package built and installed correctly
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
 python -m unittest -v test_import.py
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
 python -m unittest -v test_io.py
-
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
 # test basic avl functionality
 python -m unittest -v test_parameters.py
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
 python -m unittest -v test_analysis.py
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
 python -m unittest -v test_surf_geom.py
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
 python -m unittest -v test_contraints.py
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
 
 # tests for adjoint
 python -m unittest -v test_new_subroutines.py
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
 python -m unittest -v test_partial_derivs.py
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
 python -m unittest -v test_total_derivs.py
-
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
 # test mem ussage of pyavl and test framework
 python -m unittest -v test_tear_down.py
+free -m | awk 'NR==2{printf "Memory Usage: %s/%sMB (%.2f%%)\n", $3,$2,$3*100/$2 }'
